chore(roadExample): Move directory
