+++ jiggyjiggy/decorator/src/java/roadExample/main/Display.java
@@ -1,4 +1,4 @@
-package decorator.src.java.main;
+package decorator.src.java.roadExample.main;
 
 public abstract class Display {
     public abstract void draw();
