+++ jiggyjiggy/decorator/src/java/roadExample/main/LaneDecorator.java
@@ -1,4 +1,4 @@
-package decorator.src.java.main;
+package decorator.src.java.roadExample.main;
 
 public class LaneDecorator extends DisplayDecorator {
